hecate_pyx/requirements.txt
# Core dependencies for SQL Server connection
pyodbc>=4.0.35

# Cryptography for AES-256 encryption
cryptography>=41.0.0

# Security - 2FA/TOTP
pyotp>=2.9.0
qrcode[pil]>=7.4.2
Pillow>=10.0.0

# SSH Tunnel support
sshtunnel>=0.4.0

# GUI framework (Tkinter is included with Python)
# For future PyQt6 version (optional):
# PyQt6>=6.5.0
# PyQt6-WebEngine>=6.5.0

# Utilities
python-dateutil>=2.8.2

# Optional: for better performance monitoring
psutil>=5.9.0

# Development dependencies (optional)
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0
